mostly code style

+++ src/random.cpp
@@ -3,7 +3,7 @@
 namespace unittest {
 
 random_value<bool>::random_value()
-    : random_object<bool>(), distribution_(static_cast<int>(0), static_cast<int>(1))
+    : random_object<bool>(), distribution_(0, 1)
 {}
 
 std::shared_ptr<random_object<bool>>
